+++ bert/providers/text/single.py
@@ -1,8 +1,8 @@
 from collections import OrderedDict
-
-from arekit.common.entities.base import Entity
+from arekit.bert.providers.text.terms_mapper import StringTextTermsMapper
+from arekit.common.entities.str_mask_fmt import StringEntitiesFormatter
 from arekit.common.labels.base import Label
-from arekit.processing.text.token import Token
+from arekit.common.synonyms import SynonymsCollection
 
 
 class SingleTextProvider(object):
@@ -10,32 +10,16 @@ class SingleTextProvider(object):
     TEXT_A = u'text_a'
     TERMS_SEPARATOR = u" "
 
-    SUBJECT = u"X"
-    OBJECT = u"Y"
-    ENTITY = u"E"
-
-    def __init__(self):
-        pass
+    def __init__(self, entities_formatter, synonyms):
+        assert (isinstance(entities_formatter, StringEntitiesFormatter))
+        assert(isinstance(synonyms, SynonymsCollection))
+        self._entities_formatter = entities_formatter
+        self._synonyms = synonyms
+        self._mapper = StringTextTermsMapper(entities_formatter=entities_formatter,
+                                             synonyms=synonyms)
 
     def iter_columns(self):
         yield SingleTextProvider.TEXT_A
-
-    @staticmethod
-    def __iterate_sentence_terms(sentence_terms, s_ind, t_ind):
-
-        for i, term in enumerate(sentence_terms):
-
-            if isinstance(term, unicode):
-                yield term
-            elif isinstance(term, Entity):
-                if i == s_ind:
-                    yield SingleTextProvider.SUBJECT
-                elif i == t_ind:
-                    yield SingleTextProvider.OBJECT
-                else:
-                    yield SingleTextProvider.ENTITY
-            elif isinstance(term, Token):
-                yield term.get_original_value()
 
     @staticmethod
     def _process_text(text):
@@ -45,7 +29,9 @@ class SingleTextProvider(object):
     def add_text_in_row(self, row, sentence_terms, s_ind, t_ind, expected_label):
         assert(isinstance(row, OrderedDict))
         assert(isinstance(expected_label, Label))
-        text = self.TERMS_SEPARATOR.join(self.__iterate_sentence_terms(sentence_terms,
-                                                                       s_ind=s_ind,
-                                                                       t_ind=t_ind))
+
+        self._mapper.set_s_ind(s_ind)
+        self._mapper.set_t_ind(t_ind)
+        text = self.TERMS_SEPARATOR.join(self._mapper.iter_mapped(sentence_terms))
+
         row[self.TEXT_A] = self._process_text(text)