+++ common/entities/types.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class EntityType(Enum):
+    Object = 1
+    Subject = 2
+    SynonymSubject = 2
+    SynonymObject = 2
+    Other = 3