+++ processing/text/enums.py
@@ -1,0 +1,17 @@
+from enum import Enum
+
+
+class TermFormat(Enum):
+    """
+    Supported types of terms
+    """
+
+    """
+    Original value
+    """
+    Raw = 1
+
+    """
+    Processed by stemmer
+    """
+    Lemma = 2